+++ DemoApps/GRDBDemoiOS/GRDBDemoiOS/Person.swift
@@ -16,7 +16,7 @@ class Person : Record {
     
     // MARK: - Record
     
-    override class func databaseTableName() -> String? {
+    override class func databaseTableName() -> String {
         return "persons"
     }