+++ GRDB Tests/Public/Record/PrimaryKeyNoneTests.swift
@@ -5,7 +5,7 @@ import GRDB
 class Item: Record {
     var name: String?
     
-    override class func databaseTableName() -> String? {
+    override class func databaseTableName() -> String {
         return "items"
     }