+++ Playgrounds/JSONSynchronization.playground/Contents.swift
@@ -30,7 +30,7 @@ class Person : Record {
     var id: Int64?
     var name: String?
     
-    override class func databaseTableName() -> String? {
+    override class func databaseTableName() -> String {
         return "persons"
     }