+++ Playgrounds/Record.playground/Contents.swift
@@ -38,7 +38,7 @@ class Person : Record {
     
     // Record overrides
     
-    override class func databaseTableName() -> String? {
+    override class func databaseTableName() -> String {
         return "persons"
     }